Rebase to libffi 3.4.4.
At the request of the RISC-V developers we are rebasing from
libffi-3.4.2 to libffi-3.4.4.

In order to test the rebase we used the Fedora Packaing tools
mass-prebuild tooling to rebuild 144 packages (direct reverse
dependencies of libffi) and evaluate the risk.

There were 8 package build failures that needed analysis. Of the
failures only 1 seemed related, Chromium, which failed to build with the
new libffi. The others: bustle, llvm[9.0,10,11,12], p11-kit and ruby
failed in the A-side and B-side tests. The Chromium build timed out when
building in COPR, and then passed the build without a timeout on the
A-side test (without libffi). This failure seems spurious and so I am
moving forward with the rebase.